refactor: cross-import 적용 누락 수정
