+++ src/entities/text-parsing/lib/helpers/transform-attendance-info-to-statistic.ts
@@ -1,4 +1,4 @@
-import type { GetAttendanceOptionsAll } from '@/entities/attendance-option'
+import type { GetAttendanceOptionsAll } from '@/entities/attendance-option/@x/text-parsing'
 import type { AttendanceStatisticValue, ParsingResult } from '../../model'
 
 /**
